rename hcl parser to hcl1 for the sake of consistency
