+++ docs/examples.md
@@ -7,7 +7,7 @@ You can find examples using various other tools in the `examples` directory, inc
 * [Docker compose](https://github.com/instrumenta/conftest/tree/master/examples/compose)
 * [Dockerfile](https://github.com/instrumenta/conftest/tree/master/examples/docker)
 * [EDN](https://github.com/instrumenta/conftest/tree/master/examples/edn)
-* [HCL](https://github.com/instrumenta/conftest/tree/master/examples/hcl)
+* [HCL(HCL2)](https://github.com/instrumenta/conftest/tree/master/examples/hcl)
 * [HOCON](https://github.com/instrumenta/conftest/tree/master/examples/hocon)
 * [INI](https://github.com/instrumenta/conftest/tree/master/examples/ini)
 * [GitLab](https://github.com/instrumenta/conftest/tree/master/examples/ci)
@@ -16,7 +16,7 @@ You can find examples using various other tools in the `examples` directory, inc
 * [Multitype](https://github.com/instrumenta/conftest/tree/master/examples/multitype)
 * [Serverless Framework](https://github.com/instrumenta/conftest/tree/master/examples/serverless)
 * [Tekton](https://github.com/instrumenta/conftest/tree/master/examples/tekton)
-* [Terraform](https://github.com/instrumenta/conftest/tree/master/examples/terraform)
+* [HCL1-Terraform](https://github.com/instrumenta/conftest/tree/master/examples/hcl1)
 * [Traefik](https://github.com/instrumenta/conftest/tree/master/examples/traefik)
 * [Typescript](https://github.com/instrumenta/conftest/tree/master/examples/ts)
 * [VCL](https://github.com/instrumenta/conftest/tree/master/examples/vcl)
