+++ docs/options.md
@@ -7,7 +7,7 @@
 For the available parsers, take a look at: [parsers](parser). For instance:
 
 ```console
-$ conftest test -p examples/hcl/policy examples/hcl/terraform.tf -i hcl
+$ conftest test -p examples/hcl1/policy examples/hcl1/gke.tf -i hcl1
 FAIL - examples/hcl/terraform.tf - ALB `my-alb-listener` is using HTTP rather than HTTPS
 FAIL - examples/hcl/terraform.tf - ASG `my-rule` defines a fully open ingress
 FAIL - examples/hcl/terraform.tf - Azure disk `source` is not encrypted
