+++ examples/hcl1/gke-show.json
