+++ examples/hcl1/policy/base.rego
