+++ examples/hcl1/policy/gke.rego
