+++ examples/hcl1/policy/gke_combine.rego
@@ -1,5 +1,5 @@
 package main
 
 deny[sprintf("file path index to key value does not exist: %v", [input])] {
-    not input["examples/terraform/gke.tf"].provider[0].google[0].project == "instrumenta"
+    not input["examples/hcl1/gke.tf"].provider[0].google[0].project == "instrumenta"
 }
