+++ parser/hcl1/hcl1.go
@@ -1,4 +1,4 @@
-package hcl
+package hcl1
 
 import (
 	"fmt"
